+++ index.html
@@ -31,7 +31,7 @@
         <input type="text" id="input" placeholder="Add a Todo" />
         <button id="addTaskButton">Add Task</button>
         <div class="trashInput">
-            <img src="icons/delete.svg" job="delete" id="0" />
+          <img src="icons/delete.svg" job="delete" id="0" />
         </div>
       </div>
 
@@ -48,6 +48,7 @@
     <script type="module" src="js/markDone.js" defer></script>
     <script type="module" src="js/editTask.js" defer></script>
     <script type="module" src="js/addPagination.js" defer></script>
+
     <!-- <script type="module" src="js/filterTask.js" defer></script> -->
   </body>
 </html>
