+++ js/constants.js
@@ -21,7 +21,7 @@ export const COMPLETED_TASK = "COMPLETED_TASK";
 export const ALL_FILTER = "all";
 export const INCOMPLETE_FILTER = "incompleted";
 export const COMPLETE_FILTER = "completed";
-export const PAGINATED_NO = 9;
+export const PAGINATED_NO = 2;
 export const AFTER_BEGIN = "afterbegin";
 
 //load more, show less button
